+++ examples/with-create-react-app/src/App.tsx
@@ -1,16 +1,16 @@
-import { ConnectButton } from 'diditsdktest';
+import { ConnectButton } from '@rainbow-me/rainbowkit';
 import {
   midnightTheme,
   getDefaultWallets,
-  DiditAuthProvider,
-} from 'diditsdktest';
+  RainbowKitProvider,
+} from '@rainbow-me/rainbowkit';
 import { configureChains, createClient, WagmiConfig } from 'wagmi';
 import { mainnet, polygon, optimism, arbitrum, goerli } from 'wagmi/chains';
 import { publicProvider } from 'wagmi/providers/public';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import SecondView from './views/SecondView';
 import ThirdView from './views/thirdView';
-import { DiditProvider } from 'diditprovidertest';
+import { RainbowKitSiweNextAuthProvider } from '@rainbow-me/rainbowkit-siwe-next-auth';
 
 const { chains, provider, webSocketProvider } = configureChains(
   [
@@ -23,8 +23,11 @@ const { chains, provider, webSocketProvider } = configureChains(
   [publicProvider()]
 );
 
+const projectId = 'b0337f8e2c56c722a1fb3a4cdf893249';
+
 const { connectors } = getDefaultWallets({
   appName: 'RainbowKit demo',
+  projectId,
   chains,
 });
 
@@ -38,8 +41,8 @@ const wagmiClient = createClient({
 const App = () => {
   return (
     <WagmiConfig client={wagmiClient}>
-      <DiditProvider clientUrl="https://apx.dev.gamium.world/avatar/auth">
-        <DiditAuthProvider chains={chains} theme={midnightTheme()}>
+      <RainbowKitSiweNextAuthProvider clientUrl="https://apx.dev.gamium.world/avatar/auth">
+        <RainbowKitProvider chains={chains} theme={midnightTheme()}>
           <Router>
             <Routes>
               <Route path="/app" element={<SecondView />} />
@@ -47,8 +50,8 @@ const App = () => {
               <Route path="" element={<ConnectButton />} />
             </Routes>
           </Router>
-        </DiditAuthProvider>
-      </DiditProvider>
+        </RainbowKitProvider>
+      </RainbowKitSiweNextAuthProvider>
     </WagmiConfig>
   );
 };