+++ packages/rainbowkit-siwe-next-auth/src/index.ts
@@ -1,1 +1,1 @@
-export { DiditProvider } from './RainbowKitSiweNextAuthProvider';
+export { RainbowKitSiweNextAuthProvider } from './RainbowKitSiweNextAuthProvider';