+++ packages/rainbowkit/src/components/RainbowKitProvider/RainbowKitProvider.tsx
@@ -48,7 +48,7 @@ export type Theme =
       darkMode: ThemeVars;
     };
 
-export interface DiditAuthProviderProps {
+export interface RainbowKitProviderProps {
   chains: RainbowKitChain[];
   initialChain?: RainbowKitChain | number;
   id?: string;
@@ -67,18 +67,18 @@ export interface DiditAuthProviderProps {
 
 const defaultTheme = lightTheme();
 
-export function DiditAuthProvider({
-  chains,
-  initialChain,
-  id,
-  theme = defaultTheme,
-  children,
+export function RainbowKitProvider({
   appInfo,
-  showRecentTransactions = false,
-  coolMode = false,
   avatar,
+  chains,
+  children,
+  coolMode = false,
+  id,
+  initialChain,
   modalSize = ModalSizeOptions.WIDE,
-}: DiditAuthProviderProps) {
+  showRecentTransactions = false,
+  theme = defaultTheme,
+}: RainbowKitProviderProps) {
   usePreloadImages();
 
   useAccount({ onDisconnect: clearWalletConnectDeepLink });